transfer データオブジェクトの定義を別のファイルに移行
